requirements.txt
annotated-types==0.7.0
boto3==1.41.1
botocore==1.41.1
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
jmespath==1.0.1
nflreadpy==0.1.5
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.15.0
six==1.17.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
